+++ src/main/java/de/alpharogroup/xml/XmlToObjectExtensions.java
@@ -34,6 +34,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.thoughtworks.xstream.XStream;
 
+import de.alpharogroup.xml.factory.XStreamFactory;
 import de.alpharogroup.xml.factory.XmlMapperFactory;
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
@@ -59,7 +60,7 @@ public final class XmlToObjectExtensions
 	{
 
 		XMLDecoder dec = null;
-		T obj = null;
+		T obj;
 		try
 		{
 			final InputStream is = new ByteArrayInputStream(xmlString.getBytes());
@@ -144,33 +145,8 @@ public final class XmlToObjectExtensions
 	public static <T> T toObjectWithXStream(XStream xstream, final String xmlString,
 		final Map<String, Class<?>> aliases)
 	{
-		xstream = XmlToObjectExtensions.initializeXStream(xstream, aliases);
+		xstream = XStreamFactory.initializeXStream(xstream, aliases);
 		return (T)xstream.fromXML(xmlString);
-	}
-
-	/**
-	 * Initialize the given {@link XStream} object with the given aliases
-	 *
-	 * @param xstream
-	 *            the {@link XStream} object
-	 * @param aliases
-	 *            the aliases
-	 * @return the initialized {@link XStream} object
-	 */
-	public static XStream initializeXStream(XStream xstream, Map<String, Class<?>> aliases)
-	{
-		if (xstream == null)
-		{
-			xstream = new XStream();
-		}
-		if (aliases != null)
-		{
-			for (final Map.Entry<String, Class<?>> alias : aliases.entrySet())
-			{
-				xstream.alias(alias.getKey(), alias.getValue());
-			}
-		}
-		return xstream;
 	}
 
 	/**