+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ADDED:
  
 - new gradle.properties file created and extracted dependency versions to it
 - new XmlMapperFactory class for create XmlMapper objects
+- new XStreamFactory class for create XStream objects
 - new method in JsonStringToObjectExtensions  for conversion of json string to java objects with jackson
 - new method in ObjectToXmlExtensions for conversion of java object to xml with jackson